pyproject.toml
[tool.poetry]
name = "waffle-bot"
version = "0.1.0"
description = ""
authors = ["faith <jamienibtong@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.3.2"}
youtube-dl = "^2020.3.8"
humanize = "^2.3.0"
black = "^19.10b0"
toml = "^0.10.1"
dnspython = "^1.16.0"
SQLAlchemy = "^1.4"
alembic = "^1.5.8"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
